refactor(cmd): Extract reusable Run() function for downstream forks
Move application initialization logic from main.go into cmd/app/run.go
to allow downstream forks to reuse the core setup while adding custom
plugins via blank imports.

This enables zero-copy forks that only need a minimal main.go:
```go
package main

import (
    "homeautomation/cmd/app"
    _ "github.com/example/my-private-plugin"
)

func main() {
    app.Run()
}
```

Closes #234
